Initial project upload
Initial project upload. Some bugs or odd behaviors and results are to be expected, as there have been code revisions and cleanups that where not thoroughly tested and may have unexpected consequences.

NB: There are is file -- "lib/nlpport-0.9.x.jar" -- that is split in multipart zip files so it could to be uploaded to GitHub. You must unzip those files if you want to use it.